coco

+++ Algorithmes/A1.hpp
@@ -5,7 +5,7 @@
 
 using namespace std;
 
-int pls(const char cs, const char ct);
+unsigned short pls(const char cs, const char ct);
 string A1(string s, string t);
 
 #endif
